Description
This is a Java project I created to function as a basic calculator. The user is able to add, subtract, multiply, divide, square, and get the squareroot of any number. The user is also able to mix any expression into the input. The calculator will complete the problem in the order it was entered, not in traditional PEMDAS. For example, the user can input 2 + 5 x 7 - 4 / 5 which will come out to 9. The calculator is able to display decimal numbers, but will round any decimal value after 0.0001.